RAVTDBQ-576- resolved L1 comment

+++ core/src/main/java/org/apache/calcite/sql/fun/SqlLibraryOperators.java
@@ -1986,7 +1986,7 @@ public abstract class SqlLibraryOperators {
           SqlKind.OTHER_FUNCTION,
           ReturnTypes.INTEGER_NULLABLE, null,
           OperandTypes.family(ImmutableList.of(SqlTypeFamily.STRING, SqlTypeFamily.STRING,
-                  SqlTypeFamily.NUMERIC),
+                  SqlTypeFamily.INTEGER),
               number -> number == 2),
           SqlFunctionCategory.NUMERIC);
 }
